add support for WIT versions
This adds support for importing and exporting versioned interfaces, including
multiple versions of the "same" interface (where "same" means "has the same
identifier, modulo version").

Note that we currently include the version number in JS identifiers
unconditionally.  A future refinement would be to notice when only one version
of a given interface is imported or exported and omit the version in that case.

Fixes #60

Signed-off-by: Joel Dice <joel.dice@fermyon.com>