implement support for resources
This builds on Guy's earlier, unfinished work.  I've ported over all the
resource tests I wrote for `componentize-py` and tried to make a minimal number
of changes needed to get them passing.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

switch to upstream `js-component-bindgen`

Note that package.json still uses a relative path to pull in `jco`, and my
attempt to point it to the GitHub repo failed, so I'm not sure what the next
step is there.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

update package.json to use Jco 0.12.2

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

update `jco` dependencies

Signed-off-by: Joel Dice <joel.dice@fermyon.com>