support destructors for imported resources
This provides partial support for #68.  I've focused on imported resources to
begin with since APIs such as `wasi:http` make heavy use of child handles, which
must be disposed in the correct order to avoid traps.

Supporting destructors for exported resources may also be useful, but not as
urgent with respect to WASI.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>