feat: enable use of preexisting weval binary (#156)
* feat: enable use of pre-existing weval binary

This commit adds the ability to use a pre-existing `weval` binary as
an option to `componentize`.

This should enable builds without internet but hopefully also cut down
on failures due to being unable to fetch the binary dynamically.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* chore(docs): reword the build and cleanup workflow instructions

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* feat(dx): add a clean command for removing built StarlingMonkey

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* chore(docs): add instructions for setting up embeding

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* refactor(ci): include node versions, break up build

This commit heavily refactors the CI build pipeline to separate
different build steps to improve build times and caching.

It *should* be easier to see the dependency between the builds:

- splicer (jit/AOT independent)
- StarlingMonkey JIT
- StarlingMonkey AOT (Weval)

Along with these builds, JIT/AOT tests that should come after can pull
the relevant cached artifacts.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* refactor(ci): use top level script targets

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* feat(docs): add documentation for custom weval binary

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* chore(ci): remove pinned versions

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* fix(ci): more consistent artifact upload/download

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

---------

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>