Remove warning about stubbed imports
The warning added in #185 turns out to be too noisy: there are too many situations in which an import exists, but should be stubbed. In particular once StarlingMonkey is updated to include https://github.com/bytecodealliance/StarlingMonkey/pull/218, the warning would essentially be shown for every component, because in most cases, the target world won't contain the necessary interfaces, by design.