Make StarlingMonkey location overridable (#196)
This adds the ability to override the location the `StarlingMonkey` source is found at via the env var `STARLINGMONKEY_SRC`. This makes it much easier to develop against a local version of StarlingMonkey.

I also cleaned up the `Makefile` a bit, and moved some more stuff over into the `CMakeLists.txt` file, but that's really just minor cleanup.