Overhaul bindings generation and introduce a CLI tool (#202)
This is a pretty large patch containing a whole bunch of changes. They're not really all related, but are hard to disentangle now, after being introduced while I worked on enabling debugging of componentize-js output.

The key changes are:

## Two new CLI tools

The embedding splicer now has a native Rust CLI for both splicing and stubbing. This isn't currently all that useful, but helped debug the changes to bindings generation and splicing. I plan on building on this in the future.

Additionally, the full componentization has gained a CLI as an alternative to `jco componentize`, so running `npx componentize-js` will work and provide a reasonable interface.

## Support for unbundled JS

A key goal of the changes here was to no longer rewrite the input source code, because that messes up debugging using StarlingMonkey's debugger. As part of implementing this, we got support for loading other JS modules at initialization time "for free", so that works now, too.

## No more treeshaking

A regrettable change is that we don't support any kind of treeshaking anymore. This is much harder to implement without requiring bundled code (see the previous item). Additionally, it only worked for functions (not methods, or anything related to resources) before, and was thus of very limited value.

We can in principle, with a decent amount of effort, reintroduce at least what we had before, but I think ultimately the right approach here is to make sure that the targeted runtime environment supports the full WIT world provided during componentization.

## Small stuff

Aside from these larger items, there's a bunch of small changes, and fairly sizable refactorings.