fix: separate folder component path use
Previously, when attempting to call `componentize()` on a with a
`sourcePath` in a different directory, the call to Wizer would be
generated with the incorrect *relative* path (path prefix stripping).

This commit ensure that when Wizer is called with path prefix
stripping, the path to the mapped file is correct.